refactor(cliente): renombrar servicio token como servicio autenticacion
También se va a encargar de guardar los datos del usuario que está logueado
